Create dropdown fields
- Make `dropdownFieldJsx` which takes a key as an argument and returns a
dropdown list for either `rarity` or `sclass`. Also sets the matching
option to be selected if there is a servant currently being edited.
- Make `textFieldJsx` which returns the previously written jsx component
of a text field with the correct `placeholder` `name` and `value` fields
filled in.
- Make a component which takes in a key from `formFieldsJsx`, determines
if that key should have a text field or a dropdown field, and returns
the appropriate component